requirements.txt
Django>=4.2,<5.0
pydantic>=1.10
django-nested-admin>=3.4.0